tamper checking code
