+++ netlify/functions/upload.mjs
@@ -1,16 +1,7 @@
 import { getStore } from '@netlify/blobs'
 import { v4 as uuidv4 } from 'uuid'
-import { Web3 } from 'web3'
-import { TamperProofABI } from '../../globals';
 
 exports.handler = async (event) => {
-    const CONTRACT_ADDRESS = process.env.CONTRACT_TAMPER_PROOF
-    const contractOwnerKey = process.env.CONTRACT_OWNER_ACCOUNT
-    const web3 = new Web3(process.env.INFURA_RPC_URL)
-
-    const wallet = await web3.eth.accounts.decrypt(contractOwnerKey, "gom123#")
-    const contract = new web3.eth.Contract(TamperProofABI, CONTRACT_ADDRESS);
-
     const store = await getStore({ name: "uploads", siteID: process.env.siteID, token: process.env.TOKEN })
     if (event.httpMethod !== 'POST') {
         return { statusCode: 405, body: 'Method Not Allowed' };
@@ -18,34 +9,6 @@ exports.handler = async (event) => {
 
     const jsonBody = JSON.parse(event.body)
     const { fileName, contentType, base64File, code } = jsonBody
-
-    // TODO: file tamper verification code
-    try {
-        const toHash = `${code}/tralarelotralala/${base64File}`
-
-        const fileHash = web3.utils.sha3(toHash)
-
-        const block = await web3.eth.getBlock();
-        
-        const tx = {
-            from: wallet.address,
-            to: CONTRACT_ADDRESS,
-            maxFeePerGas: block.baseFeePerGas * 2n,
-            maxPriorityFeePerGas: 1,
-            data: contract.methods.uploadFileHash(fileHash).encodeABI()
-        }
-
-        const signtx = await web3.eth.accounts.signTransaction(tx, wallet.privateKey)
-        
-        console.log("signed a transaction")
-        
-        await web3.eth.sendSignedTransaction(signtx.rawTransaction)
-    
-        console.log("successfully sent signed transaction?")
-    } catch (e) {
-        console.log(e)
-    }
-    // TODO: file tamper verification code end
     
     const fileBuffer = Buffer.from(base64File, 'base64'); // Lambda sends body as base64
     const blobId = uuidv4()
